Use consistent EOL (Unix) no matter what OS is used to generate
Before, EOL was dynamic, based on the system we run the generation on - this was highly undesirable. We do development on Windows and Unix systems, so the EOL alone would result in changes in generated translation files even without any actual changes. This matters if the translation files are revisioned.